cleaning up api for better working
