+++ src/app/api/[userId]/projects/route.js
@@ -1,62 +1,74 @@
-import { NextRequest, NextResponse } from "next/server";
-import connectDB from "@/app/db/connectDB";
-import Project from "@/app/db/model/Project";
-import User from "@/app/db/model/User";
-import Employee from "@/app/db/model/Employee";
-import Login from "@/app/components/Login";
+// import { NextRequest, NextResponse } from "next/server";
+// import connectDB from "@/app/db/connectDB";
+// import Project from "@/app/db/model/Project";
+// import User from "@/app/db/model/User";
+// import Employee from "@/app/db/model/Employee";
+// import Login from "@/app/components/Login";
 
-export async function POST(request, { params, searchParams }, response) {
-  await connectDB();
-  const data = await request.json();
-  const method = await data.method;
-  const { userId, empId } = params;
-  let proId = params.proId;
+// export async function POST(request, { params, searchParams }, response) {
+//   await connectDB();
+//   const data = await request.json();
+//   const method = await data.method;
+//   const { userId, empId } = params;
+//   let proId = params.proId;
 
-  console.log("params server", params);
-  console.log("method ", method);
+//   console.log("params server", params);
+//   console.log("method ", method);
 
-  if (method === "CREATEPROJECT") {
-    try {
-      const newProject = await Project.create(data);
-      await User.findByIdAndUpdate(
-        userId,
-        { $push: { projects: newProject._id } },
-        { new: true }
-      );
-      console.log("Server side project created");
-      return NextResponse.json({ status: 201 });
-    } catch (error) {
-      console.log("Error posting new project", error);
-      return NextResponse.json({ status: 400 });
-    }
-  } else if (method === "DELETEEMPLOYEE") {
-    try {
-      const employeeToDelete = await Employee.findByIdAndDelete(empId);
-      await Project.findByIdAndUpdate(proId, {
-        $pull: { employees: empId },
-      });
-      return NextResponse.json({ status: 201 });
-    } catch (error) {
-      console.log("Error deleting employee", error);
-      return NextResponse.json({ status: 400 });
-    }
-  } else if (method === "CREATEEMPLOYEE") {
-    proId = data.project;
+  // if (method === "CREATEPROJECT") {
+  //   try {
+  //     const newProject = await Project.create(data);
+  //     await User.findByIdAndUpdate(
+  //       userId,
+  //       { $push: { projects: newProject._id } },
+  //       { new: true }
+  //     );
+  //     console.log("Server side project created");
+  //     return NextResponse.json({ status: 201 });
+  //   } catch (error) {
+  //     console.log("Error posting new project", error);
+  //     return NextResponse.json({ status: 400 });
+  //   }
+  // }
+  // if (method === "DELETEEMPLOYEE") {
+  //   try {
+  //     const employeeToDelete = await Employee.findByIdAndDelete(empId);
+  //     await Project.findByIdAndUpdate(proId, {
+  //       $pull: { employees: empId },
+  //     });
+  //     return NextResponse.json({ status: 201 });
+  //   } catch (error) {
+  //     console.log("Error deleting employee", error);
+  //     return NextResponse.json({ status: 400 });
+  //   }
+  // } else if (method === "CREATEEMPLOYEE") {
+  //   proId = data.project;
 
-    try {
-      const newEmployee = await Employee.create(data);
-      await Project.findByIdAndUpdate(
-        proId,
-        { $push: { employees: newEmployee._id } },
-        { new: true }
-      );
+  //   try {
+  //     const newEmployee = await Employee.create(data);
+  //     await Project.findByIdAndUpdate(
+  //       proId,
+  //       { $push: { employees: newEmployee._id } },
+  //       { new: true }
+  //     );
 
-      return NextResponse.json({ status: 201 });
-    } catch (error) {
-      console.log("Error creating new employee", error);
-      return NextResponse.json({ status: 400 });
-    }
-  }
+  //     return NextResponse.json({ status: 201 });
+  //   } catch (error) {
+  //     console.log("Error creating new employee", error);
+  //     return NextResponse.json({ status: 400 });
+  //   }
+  // } else if (method === "EDITEMPLOYEE") {
+  //   // proId = data.project;
+
+  //   try {
+  //     const updatedEmployee = await Employee.findByIdAndUpdate(empId, data);
+
+  //     return NextResponse.json({ status: 201 });
+  //   } catch (error) {
+  //     console.log("Error editing employee", error);
+  //     return NextResponse.json({ status: 400 });
+  //   }
+  // }
 }
 
 /*
